# rl/plugins/VIM/kubernetes/mtp_plugin_kubernetes/swagger/swagger.yaml
...
swagger: "2.0"
...
info:
  description: "MTP Manager API"
  version: "1.0"
  title: "MTP Manager API"
  termsOfService: "/tos"
paths:
...
  /network_resources:
    get:
      tags:
      - "virtualisedNetworkResources"
      operationId: "vi_mquery_networks"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "networkQueryFilter"
        description: "Query filter based on e.g. name, identifier, meta-data information\
          \ or status information, expressing the type of information to be retrieved.\
          \ It can also be used to specify one or more resources to be queried by\
          \ providing their identifiers."
        required: true
        schema:
          $ref: "#/definitions/Filter"
      responses:
        200:
          description: "Element containing information about the virtual network resource(s)\
            \ matching the filter. The cardinality can be 0 if no matching network\
            \ resources exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VirtualNetwork"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_controller"
    post:
      tags:
      - "virtualisedNetworkResources"
      operationId: "vi_mallocate_network"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "params"
        required: true
        schema:
          $ref: "#/definitions/AllocateNetworkRequest"
      responses:
        201:
          description: "Identifier of the created Compute Flavour."
          schema:
            $ref: "#/definitions/AllocateNetworkResult"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        409:
          description: "network already added"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_controller"
    delete:
      tags:
      - "virtualisedNetworkResources"
      operationId: "vi_mterminate_networks"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "networkResourceId"
        in: "query"
        description: "Identifier of the virtualised network resource(s) to be terminated."
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "Identifier(s) of the virtualised network resource(s) successfully\
            \ terminated."
          schema:
            type: "array"
            items:
              type: "string"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_controller"
  /network_resources/capacities:
    get:
      tags:
      - "virtualisedNetworkResources"
      operationId: "query_network_capacity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "QueryNetworkCapacityRequest"
        required: true
        schema:
          $ref: "#/definitions/QueryNetworkCapacityRequest"
      responses:
        200:
          description: "The capacity during the requested time period. The scope is\
            \ according to parameter zoneId of the request during the time interval."
          schema:
            $ref: "#/definitions/CapacityInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_controller"
  /network_resources/nfvi-pop-network-information:
    get:
      tags:
      - "virtualisedNetworkResources"
      operationId: "query_nfvi_po_p_network_information"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "NfviPopNetworkInformationRequest"
        description: "Input filter for selecting information to query."
        required: true
        schema:
          $ref: "#/definitions/Filter"
      responses:
        200:
          description: "The filtered information that has been retrieved. The cardinality\
            \ can be 0 if no matching information exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/NfviPop"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.virtualised_network_resources_controller"
...
  /software_images:
    get:
      tags:
      - "softwareImages"
      operationId: "query_software_images"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "The information of all software images matching the query."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SoftwareImageInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.software_images_controller"
    post:
      tags:
      - "softwareImages"
      operationId: "add_software_image"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SoftwareImageAddQuery"
      responses:
        201:
          description: "Metadata about the Software Image that has been added."
          schema:
            $ref: "#/definitions/SoftwareImageInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        409:
          description: "software image already added"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.software_images_controller"
  /software_images/{id}:
    get:
      tags:
      - "softwareImages"
      operationId: "query_software_image"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The identifier of the software image to be queried."
        required: true
        type: "string"
      responses:
        200:
          description: "The information of the software image matching the query."
          schema:
            $ref: "#/definitions/SoftwareImageInformation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.software_images_controller"
    delete:
      tags:
      - "softwareImages"
      operationId: "delete_software_image"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The identifier of the software image to be deleted."
        required: true
        type: "string"
      responses:
        200:
          description: "The identifier of the software image successfully deleted."
          schema:
            type: "string"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "mtp_plugin_kubernetes.controllers.software_images_controller"
definitions:
...
  AllocateNetworkRequest:
    type: "object"
    required:
    - "affinityOrAntiAffinityConstraints"
    - "locationConstraints"
    - "metadata"
    - "networkResourceName"
    - "networkResourceType"
    - "reservationId"
    - "resourceGroupId"
    - "typeNetworkData"
    - "typeNetworkPortData"
    - "typeSubnetData"
    properties:
      affinityOrAntiAffinityConstraints:
        type: "string"
        description: "The binary software image file."
      locationConstraints:
        type: "string"
        description: "Controls the visibility of the image. In case of \"private\"\
          \ value the image is available only for the tenant assigned to the provided\
          \ resourceGroupId and the administrator tenants of the VIM while in case\
          \ of \"public\" value, all tenants of the VIM can use the image."
      metadata:
        type: "array"
        description: "The binary software image file."
        items:
          $ref: "#/definitions/metaData_inner"
      networkResourceName:
        type: "string"
        description: "Name provided by the consumer for the virtualised network resource\
          \ to allocate. It can be used for identifying resources from consumer side."
      networkResourceType:
        type: "string"
        description: "Type of virtualised network resource. Possible values are: \"\
          network\", \"subnet\" or network-port."
      reservationId:
        type: "string"
        description: "Identifier of the resource reservation applicable to this virtualised\
          \ resource management operation."
      resourceGroupId:
        type: "string"
        description: "Unique identifier of the \"infrastructure resource group\",\
          \ logical grouping of virtual resources assigned to a tenant within an Infrastructure\
          \ Domain."
      typeNetworkData:
        type: "string"
        description: "The network data provides information about the particular virtual\
          \ network resource to create. Cardinality can be \"0\" depending on the\
          \ value of networkResourceType."
      typeNetworkPortData:
        type: "string"
        description: "The binary software image file."
      typeSubnetData:
        type: "string"
        description: "The binary software image file."
    example:
      networkResourceType: "networkResourceType"
      metadata:
      - value: "value"
        key: "key"
      - value: "value"
        key: "key"
      typeSubnetData: "typeSubnetData"
      networkResourceName: "networkResourceName"
      affinityOrAntiAffinityConstraints: "affinityOrAntiAffinityConstraints"
      reservationId: "reservationId"
      typeNetworkPortData: "typeNetworkPortData"
      locationConstraints: "locationConstraints"
      resourceGroupId: "resourceGroupId"
      typeNetworkData: "typeNetworkData"
  AllocateNetworkResult:
    type: "object"
    required:
    - "networkData"
    - "networkPortData"
    - "subnetData"
    properties:
      networkData:
        $ref: "#/definitions/AllocateNetworkResult_networkData"
      networkPortData:
        $ref: "#/definitions/AllocateNetworkResult_networkPortData"
      subnetData:
        $ref: "#/definitions/AllocateNetworkResult_subnetData"
    example:
      networkData:
        subnet: "subnet"
        networkResourceName: "networkResourceName"
        bandwidth: 0.80082819046101150206595775671303272247314453125
        networkQoS:
        - qosName: "qosName"
          qosValue: "qosValue"
        - qosName: "qosName"
          qosValue: "qosValue"
        zoneId: "zoneId"
        networkResourceId: "networkResourceId"
        operationalState: "operationalState"
        networkType: "networkType"
        networkPort:
        - portType: "portType"
          metadata:
          - value: "value"
            key: "key"
          - value: "value"
            key: "key"
          resourceId: "resourceId"
          bandwidth: 6.02745618307040320615897144307382404804229736328125
          segmentId: "segmentId"
          networkId: "networkId"
          operationalState: "operationalState"
          attachedResourceId: "attachedResourceId"
        - portType: "portType"
          metadata:
          - value: "value"
            key: "key"
          - value: "value"
            key: "key"
          resourceId: "resourceId"
          bandwidth: 6.02745618307040320615897144307382404804229736328125
          segmentId: "segmentId"
          networkId: "networkId"
          operationalState: "operationalState"
          attachedResourceId: "attachedResourceId"
        isShared: true
        segmentType: "segmentType"
        sharingCriteria: "sharingCriteria"
      subnetData:
        metadata:
        - value: "value"
          key: "key"
        - value: "value"
          key: "key"
        resourceId: "resourceId"
        ipVersion: "ipVersion"
        isDhcpEnabled: true
        addressPool: "addressPool"
        cidr: "cidr"
        gatewayIp: "gatewayIp"
        networkId: "networkId"
        operationalState: "operationalState"
      networkPortData:
        portType: "portType"
        metadata:
        - value: "value"
          key: "key"
        - value: "value"
          key: "key"
        resourceId: "resourceId"
        bandwidth: 6.02745618307040320615897144307382404804229736328125
        segmentId: "segmentId"
        networkId: "networkId"
        operationalState: "operationalState"
        attachedResourceId: "attachedResourceId"
  CapacityInformation:
    type: "object"
    required:
    - "allocatedCapacity"
    - "availableCapacity"
    - "reservedCapacity"
    - "totalCapacity"
    properties:
      allocatedCapacity:
        type: "string"
        description: "The allocated capacity is usually specified as the current allocated\
          \ capacity."
      availableCapacity:
        type: "string"
        description: "The free capacity available for allocation and reservation.\
          \ It can be specified in terms of current capacity; or minimum and maximum\
          \ capacity; average capacity; or other statistical measurement in the specified\
          \ time interval. The set of measurements is to be defined during Stage 3."
      reservedCapacity:
        type: "string"
        description: "The reserved capacity. It can be specified in terms of current\
          \ capacity; or minimum and maximum capacity; average capacity; or other\
          \ statistical measurement in the specified time interval. The set of measurements\
          \ is to be defined during Stage 3."
      totalCapacity:
        type: "string"
        description: "The total capacity is usually specified as a fixed capacity\
          \ without variations in time. The set of measurements is left to Stage 3."
    example:
      availableCapacity: "availableCapacity"
      reservedCapacity: "reservedCapacity"
      allocatedCapacity: "allocatedCapacity"
      totalCapacity: "totalCapacity"
...
  Filter:
    type: "object"
    required:
    - "filter"
    properties:
      filter:
        type: "object"
        description: "The filter is used to select elements for query purpose."
        properties: {}
    example:
      filter: "{}"
  NfviPop:
    type: "object"
    required:
    - "geographicalLocationInfo"
    - "networkConnectivityEndpoint"
    - "nfviPopId"
    - "vimId"
    properties:
      geographicalLocationInfo:
        type: "string"
        description: "It provides information about the geographic location (e.g.\
          \ geographic coordinates or address of the building, etc.) of the NFVI resources\
          \ that the VIM manages."
      networkConnectivityEndpoint:
        type: "string"
        description: "Information about network connectivity endpoints to the NFVI-PoP\
          \ that the VIM manages which helps build topology information relative to\
          \ NFVI-PoP connectivity to other NFVI-PoP or N-PoP."
      nfviPopId:
        type: "string"
        description: "Identification of the NFVI-PoP."
      vimId:
        type: "string"
        description: "Identification of the VIM."
    example:
      networkConnectivityEndpoint: "networkConnectivityEndpoint"
      vimId: "vimId"
      nfviPopId: "nfviPopId"
      geographicalLocationInfo: "geographicalLocationInfo"
...
  QueryNetworkCapacityRequest:
    type: "object"
    required:
    - "attributeSelector"
    - "networkResourceTypeId"
    - "resourceCriteria"
    - "timePeriod"
    - "zoneId"
    properties:
      attributeSelector:
        type: "string"
        description: "Input parameter for selecting which capacity information (i.e.\
          \ available, total, reserved and/or allocated capacity) is queried. When\
          \ not present, all four values are requested."
      networkResourceTypeId:
        type: "string"
        description: "Identifier of the resource type for which the issuer wants to\
          \ know the available, total, reserved and/or allocated capacity."
      resourceCriteria:
        type: "string"
        description: "Input capacity computation parameter for selecting the characteristics\
          \ of the virtual network for which the issuer wants to know the available,\
          \ total, reserved and/or allocated capacity. Selecting parameters/attributes\
          \ that shall be used are defined in the VirtualNetworkResourceInformation\
          \ information element. This information element and the networkResourceTypeID\
          \ are mutually exclusive."
      timePeriod:
        $ref: "#/definitions/QueryComputeCapacityRequest_timePeriod"
      zoneId:
        type: "string"
        description: "When specified this parameter identifies the resource zone for\
          \ which the capacity is requested. When not specified the total capacity\
          \ managed by the VIM instance will be returned."
    example:
      networkResourceTypeId: "networkResourceTypeId"
      resourceCriteria: "resourceCriteria"
      timePeriod:
        startTime: "2000-01-23T04:56:07.000+00:00"
        stopTime: "2000-01-23T04:56:07.000+00:00"
      zoneId: "zoneId"
      attributeSelector: "attributeSelector"
...
  SoftwareImageAddQuery:
    type: "object"
    required:
    - "name"
    - "provider"
    - "resourceGroupId"
    - "softwareImage"
    - "userMetadata"
    - "version"
    - "visibility"
    properties:
      name:
        type: "string"
        description: "The name of the software image."
      provider:
        type: "string"
        description: "The provider of the software image."
      resourceGroupId:
        type: "string"
        description: "Unique identifier of the \"infrastructure resource group\",\
          \ logical grouping of virtual resources assigned to a tenant within an Infrastructure\
          \ Domain."
      softwareImage:
        type: "string"
        description: "The binary software image file."
      userMetadata:
        type: "array"
        description: "User-defined metadata."
        items:
          $ref: "#/definitions/metaData_inner"
      version:
        type: "string"
        description: "The version of the software image."
      visibility:
        type: "string"
        description: "Controls the visibility of the image. In case of \"private\"\
          \ value the image is available only for the tenant assigned to the provided\
          \ resourceGroupId and the administrator tenants of the VIM while in case\
          \ of \"public\" value, all tenants of the VIM can use the image."
    example:
      visibility: "visibility"
      provider: "provider"
      softwareImage: "softwareImage"
      name: "name"
      resourceGroupId: "resourceGroupId"
      userMetadata:
      - value: "value"
        key: "key"
      - value: "value"
        key: "key"
      version: "version"
  SoftwareImageInformation:
    type: "object"
    required:
    - "checksum"
    - "containerFormat"
    - "createdAt"
    - "diskFormat"
    - "id"
    - "minDisk"
    - "minRam"
    - "name"
    - "provider"
    - "size"
    - "status"
    - "updatedAt"
    - "version"
    properties:
      checksum:
        type: "string"
        description: "The checksum of the software image file."
      containerFormat:
        type: "string"
        description: "The container format indicates whether the software image is\
          \ in a file format that also contains metadata about the actual software."
      createdAt:
        type: "string"
        description: "The created time of this software image."
      diskFormat:
        type: "string"
        description: "The disk format of a software image is the format of the underlying\
          \ disk image."
      id:
        type: "string"
        description: "The identifier of this software image."
      minDisk:
        type: "string"
        description: "The minimal Disk for this software image."
      minRam:
        type: "string"
        description: "The minimal RAM for this software image."
      name:
        type: "string"
        description: "The name of this software image."
      provider:
        type: "string"
        description: "The provider of this software image."
      size:
        type: "string"
        description: "The size of this software image."
      status:
        type: "string"
        description: "The status of this software image."
      updatedAt:
        type: "string"
        description: "The updated time of this software image."
        x-nullable: true
      version:
        type: "string"
        description: "The version of this software image."
        x-nullable: true
    example:
      minDisk: "minDisk"
      version: "version"
      createdAt: "createdAt"
      size: "size"
      provider: "provider"
      minRam: "minRam"
      checksum: "checksum"
      name: "name"
      containerFormat: "containerFormat"
      id: "id"
      diskFormat: "diskFormat"
      status: "status"
      updatedAt: "updatedAt"
...
  VirtualNetwork:
    type: "object"
    required:
    - "bandwidth"
    - "isShared"
    - "networkPort"
    - "networkQoS"
    - "networkResourceId"
    - "networkResourceName"
    - "networkType"
    - "operationalState"
    - "segmentType"
    - "sharingCriteria"
    - "subnet"
    - "zoneId"
    properties:
      bandwidth:
        type: "number"
        description: "Minimum network bandwidth (in Mbps)."
      isShared:
        type: "boolean"
        description: "It defines whether the virtualised network is shared among consumers."
      networkPort:
        type: "array"
        description: "Element providing information of an instantiated virtual network\
          \ port."
        items:
          $ref: "#/definitions/AllocateNetworkResult_networkData_networkPort"
      networkQoS:
        type: "array"
        description: "Element providing information about Quality of Service attributes\
          \ that the network shall support. The cardinality can be \"0\" to allow\
          \ for networks without any specified QoS requirements."
        items:
          $ref: "#/definitions/AllocateNetworkResult_networkData_networkQoS"
      networkResourceId:
        type: "string"
        description: "Identifier of the virtualised network resource."
      networkResourceName:
        type: "string"
        description: "Name of the virtualised network resource."
      networkType:
        type: "string"
        description: "The type of network that maps to the virtualised network. This\
          \ list is extensible. Examples are: \"local\", \"vlan\", \"vxlan\", \"gre\"\
          , \"l3-vpn\", etc. The cardinality can be \"0\" to cover the case where\
          \ this attribute is not required to create the virtualised network."
      operationalState:
        type: "string"
        description: "The operational state of the virtualised network."
      segmentType:
        type: "string"
        description: "The isolated segment for the virtualised network. For instance,\
          \ for a \"vlan\" networkType, it corresponds to the vlan identifier; and\
          \ for a \"gre\" networkType, this corresponds to a gre key. The cardinality\
          \ can be \"0\" to allow for flat networks without any specific segmentation."
      sharingCriteria:
        type: "string"
        description: "Only present for shared networks. Indicate the sharing criteria\
          \ for this network. This criteria might be a list of authorized consumers."
      subnet:
        type: "string"
        description: "Only present if the network provides layer 3 connectivity."
      zoneId:
        type: "string"
        description: "If present, it identifies the Resource Zone where the virtual\
          \ network resources have been allocated."
    example:
      subnet: "subnet"
      networkResourceName: "networkResourceName"
      bandwidth: 0.80082819046101150206595775671303272247314453125
      networkQoS:
      - qosName: "qosName"
        qosValue: "qosValue"
      - qosName: "qosName"
        qosValue: "qosValue"
      zoneId: "zoneId"
      networkResourceId: "networkResourceId"
      operationalState: "operationalState"
      networkType: "networkType"
      networkPort:
      - portType: "portType"
        metadata:
        - value: "value"
          key: "key"
        - value: "value"
          key: "key"
        resourceId: "resourceId"
        bandwidth: 6.02745618307040320615897144307382404804229736328125
        segmentId: "segmentId"
        networkId: "networkId"
        operationalState: "operationalState"
        attachedResourceId: "attachedResourceId"
      - portType: "portType"
        metadata:
        - value: "value"
          key: "key"
        - value: "value"
          key: "key"
        resourceId: "resourceId"
        bandwidth: 6.02745618307040320615897144307382404804229736328125
        segmentId: "segmentId"
        networkId: "networkId"
        operationalState: "operationalState"
        attachedResourceId: "attachedResourceId"
      isShared: true
      segmentType: "segmentType"
      sharingCriteria: "sharingCriteria"
...